+++ plugins/hf-ui.js
@@ -10,7 +10,8 @@ Vue.use(hfui);
 // 设置处理函数
 hfui.setUtils({
 	async uploadFile({ path, name }) {
-		await delay(1000);
+		// throw new Error('上传文件失败');
+		await delay(3000);
 		return {
 			url: path,	// 图片可访问绝对路径
 			path: path	// 相对路径(保存)